dwarves: Enlarge vtable_entry
Since we have two bytes after it in a hole, use them for vtable_entry,
hell knows if there isn't any crazy code with that many vtable
entries...

+++ dwarves.h
@@ -708,7 +708,7 @@ struct function {
 	uint8_t		 external:1;
 	uint8_t		 accessibility:2; /* DW_ACCESS_{public,protected,private} */
 	uint8_t		 virtuality:2; /* DW_VIRTUALITY_{none,virtual,pure_virtual} */
-	int16_t		 vtable_entry;
+	int32_t		 vtable_entry;
 	struct list_head vtable_node;
 	/* fields used by tools */
 	union {
